Refactoring; Add TestOptimizeLinear
