+++ cleverhans/future/torch/attacks/__init__.py
@@ -1,3 +1,4 @@
+# pylint: disable=missing-docstring
 from cleverhans.future.torch.attacks.attack import *
 from cleverhans.future.torch.attacks.projected_gradient_descent import *
 from cleverhans.future.torch.attacks.fast_gradient_method import *
